Temporary checkin for FileDescriptors
...and support of FileDescriptors.  This is temporary
because while we need this right away in order to support
queries, we do not want to do this in the long run.  Impala
uses catalogd to store this information rather than go to the
file system on every query.  Going to the file system on every
query is...um...probably too slow.

Change-Id: Icba1ce198969b0fa8c12399dbccf25ba88365896